Fix for neutralize target/rescue operative missions
Given an ini-configurable list of "important" character template names,
if any soldier eligible for evac is carrying a unit of these names they
will be evacuated first, and all other units will be done in a follow-up
game state triggered by an event after the first one is processed. This
gives kismet a chance to process the VIP being evacuated while other
units are still on the board.

This would break the visualization into two chunks, so while some units
may not actually be evacuating in the state, visualize their evacs
anyway, and complete the evac state change with no visualization later.