Add check for evac ability availability
The Evac ability is listed as not being always available
(bAllowedByDefault=false) and is specifically enabled in certain mission
scripts.

Evac All isn't enabled in those mission scripts, so we need to make it
always available. To ensure it doesn't show up in the action bar when
evac is disabled by a mission script we need a new custom condition that
tests if the Evac ability is globally disabled.

This prevents the evac all icon from showing up on the ability bar on
missions such as Lost and Abandoned. It doesn't do anything when used
since the Evac ability isn't enabled, but it clutters the bar.