Skip building no-evac tile list for ADVENT evac zones.
A user provided a save in a neutralize field commander mission where
the game crashed after the advent evac zone spawned. In the saved game
it spawned along the edge of the map where some tiles are blocked. The
no-evac tiles show as highlighted while the timer is still counting down
and only the flare is present, and then on the final turn when the evac
zone itself spawns the game crashes.

This mission spawns two evac zones, the first one when advent become
aware of XCOM. This uses the normal SeqAct_SpawnEvacZone kismet action
with the team set to Alien and a blueprint override to show only the
flare and not a full evac zone. The timer starts counting down. When it
reaches zero a new evac zone is spawned over top of the old one, this
time with the normal evac zone blueprint and the AI is updated to get
the field commander to consider evac.

Having two evac zones confuses evac all and causes a crash. Since the
flare is actually a full evac zone if it has some blocked tiles these
will show up even before the evac zone actually pops, which is also
strange. Also reproduced the crash when using XCOM evac zones if one
is thrown with some blocked tiles and then a new one is placed, also
with blocked tiles. The fix is to adjust how the no evac tiles are
cleaned up. When a zone state object is re-used, also reuse the same
visualizer instead of making a new one.

Also disabled no-evac tiles altogether on ADVENT zones, evac all only
cares about xcom zones.